requirements.txt
altair
numpy
pandas
pydeck
streamlit
folium
branca
Jinja2
Requests
streamlit_folium 
streamlit-pills
snowflake-connector-python
snowflake-connector
Faker